Move header files to separate directory
This way the include paths inside the project are similar to the ones a
user would use, and building examples with CMake from inside this
projects will be possible, if that is desired in the future.
