+++ include/libcgi/config.h.in
