+++ include/libcgi/error.h
