+++ include/libcgi/session.h
