+++ src/general.c
@@ -23,8 +23,8 @@
 #include <string.h>
 #include <stdlib.h>
 
-#include "error.h"
-#include "cgi.h"
+#include "libcgi/cgi.h"
+#include "libcgi/error.h"
 
 struct iso8859_15 {
 	char code;
