+++ src/list.c
@@ -25,8 +25,8 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "error.h"
-#include "cgi.h"
+#include "libcgi/cgi.h"
+#include "libcgi/error.h"
 
 // Add a new item to the list
 void slist_add(formvars *item, formvars **start, formvars **last)
